Rename queryUtil to urlQuery
